+++ pylops/basicoperators/FirstDerivative.py
@@ -34,7 +34,7 @@ class FirstDerivative(LinearOperator):
         Sampling step :math:`\Delta x`.
     edge : :obj:`bool`, optional
         Use reduced order derivative at edges (``True``) or
-        ignore them (``False``)
+        ignore them (``False``) for centered derivative
     dtype : :obj:`str`, optional
         Type of elements in input array.
     kind : :obj:`str`, optional