+++ pylops/basicoperators/SecondDerivative.py
@@ -34,9 +34,11 @@ class SecondDerivative(LinearOperator):
         Sampling step :math:`\Delta x`.
     edge : :obj:`bool`, optional
         Use reduced order derivative at edges (``True``) or
-        ignore them (``False``)
+        ignore them (``False``) for centered derivative
     dtype : :obj:`str`, optional
         Type of elements in input array.
+    kind : :obj:`str`, optional
+        Derivative kind (``forward``, ``centered``, or ``backward``).
 
     Attributes
     ----------
@@ -57,10 +59,28 @@ class SecondDerivative(LinearOperator):
     .. math::
         y[i] = (x[i+1] - 2x[i] + x[i-1]) / \Delta x^2
 
+    while the second-order forward stencil is:
+
+    .. math::
+        y[i] = (x[i+2] - 2x[i+1] + x[i]) / \Delta x^2
+
+    and the second-order backward stencil is:
+
+    .. math::
+        y[i] = (x[i] - 2x[i-1] + x[i-2]) / \Delta x^2
+
     """
 
     def __init__(
-        self, N, dims=None, axis=-1, dir=None, sampling=1, edge=False, dtype="float64"
+        self,
+        N,
+        dims=None,
+        axis=-1,
+        dir=None,
+        sampling=1,
+        edge=False,
+        dtype="float64",
+        kind="centered",
     ):
         self.N = N
         self.sampling = sampling
@@ -84,11 +104,64 @@ class SecondDerivative(LinearOperator):
         else:
             axis = axis
         self.axis = normalize_axis_index(axis, len(self.dims))
+        self.kind = kind
         self.shape = (self.N, self.N)
         self.dtype = np.dtype(dtype)
         self.explicit = False
 
-    def _matvec(self, x):
+        # choose _matvec and _rmatvec kind
+        if self.kind == "forward":
+            self._matvec = self._matvec_forward
+            self._rmatvec = self._rmatvec_forward
+        elif self.kind == "centered":
+            self._matvec = self._matvec_centered
+            self._rmatvec = self._rmatvec_centered
+        elif self.kind == "backward":
+            raise NotImplementedError(
+                "backward is not yet available, use forward or centered"
+            )
+        else:
+            raise NotImplementedError("kind must be forward, centered, or backward")
+
+    def _matvec_forward(self, x):
+        ncp = get_array_module(x)
+        if not self.reshape:
+            x = x.squeeze()
+            y = ncp.zeros(self.N, self.dtype)
+            y[:-2] = (x[2:] - 2 * x[1:-1] + x[0:-2]) / self.sampling ** 2
+        else:
+            x = ncp.reshape(x, self.dims)
+            if self.axis > 0:  # need to bring the dim. to derive to first dim.
+                x = ncp.swapaxes(x, self.axis, 0)
+            y = ncp.zeros(x.shape, self.dtype)
+            y[:-2] = (x[2:] - 2 * x[1:-1] + x[0:-2]) / self.sampling ** 2
+            if self.axis > 0:
+                y = ncp.swapaxes(y, 0, self.axis)
+            y = y.ravel()
+        return y
+
+    def _rmatvec_forward(self, x):
+        ncp = get_array_module(x)
+        if not self.reshape:
+            x = x.squeeze()
+            y = ncp.zeros(self.N, self.dtype)
+            y[0:-2] += (x[:-2]) / self.sampling ** 2
+            y[1:-1] -= (2 * x[:-2]) / self.sampling ** 2
+            y[2:] += (x[:-2]) / self.sampling ** 2
+        else:
+            x = ncp.reshape(x, self.dims)
+            if self.axis > 0:  # need to bring the dim. to derive to first dim.
+                x = ncp.swapaxes(x, self.axis, 0)
+            y = ncp.zeros(x.shape, self.dtype)
+            y[0:-2] += (x[:-2]) / self.sampling ** 2
+            y[1:-1] -= (2 * x[:-2]) / self.sampling ** 2
+            y[2:] += (x[:-2]) / self.sampling ** 2
+            if self.axis > 0:
+                y = ncp.swapaxes(y, 0, self.axis)
+            y = y.ravel()
+        return y
+
+    def _matvec_centered(self, x):
         ncp = get_array_module(x)
         if not self.reshape:
             x = x.squeeze()
@@ -111,7 +184,7 @@ class SecondDerivative(LinearOperator):
             y = y.ravel()
         return y
 
-    def _rmatvec(self, x):
+    def _rmatvec_centered(self, x):
         ncp = get_array_module(x)
         if not self.reshape:
             x = x.squeeze()